+++ CHANGELOG.md
@@ -88,3 +88,4 @@
 # 0.1.9
 * Change `Stack` definition: from type synonymous to newtype, change operations accordingly
 * Define `.:.` composition: the same thing but with reverse order
+* Change `Nonempty` type family, family instances should be defined for each data structure